+++ sugar/lib/crdt.dart
@@ -1,1 +1,2 @@
+// TODO: add documentation.
 export 'src/crdt/string_index.dart';